ams-qd:  Fix non-ASCII text width measurement
This fixes the regression from 9fd63c247284463db503258f1d879ba080aa205e,
which fixed the overflow when lastChar is 255, but caused a new overflow
when measuring or drawing non-ASCII characters, i.e. those with the high
bit set, which are negative in signed operations.  Use unsigned instead.

+++ 68k/modules/ams-qd/Text.cc
@@ -145,7 +145,7 @@ pascal short StdTxMeas_patch( short        n,
 	
 	while ( --n >= 0 )
 	{
-		short c = *p++ - rec.firstChar;
+		short c = (uint8_t) *p++ - rec.firstChar;
 		
 		if ( c > missingChar )
 		{
@@ -309,7 +309,7 @@ pascal void StdText_patch( short n, const char* p, Point numer, Point denom )
 	
 	while ( --n >= 0 )
 	{
-		short c = *p++ - rec.firstChar;
+		short c = (uint8_t) *p++ - rec.firstChar;
 		
 		if ( c > missingChar )
 		{
@@ -395,7 +395,7 @@ pascal void MeasureText_patch( short count, const char* p, short* locs )
 	
 	while ( --count >= 0 )
 	{
-		short c = *p++ - rec.firstChar;
+		short c = (uint8_t) *p++ - rec.firstChar;
 		
 		if ( c == '\r' )
 		{
